rafter service and dao finished!

+++ src/com/river/persistence/dao/implementation/RafterDAOImpl.java
@@ -76,15 +76,19 @@ public class RafterDAOImpl extends AbstractDAO implements RafterDAO{
 	}
 
 	public Rafter delete(Rafter toDelete) {
-		Session session = sessionFactory.getCurrentSession();
-		Transaction tx = session.beginTransaction();
-		try{
-			session.delete(toDelete);
-			tx.commit();
-		}catch(HibernateException e){
-			e.printStackTrace();
+		if (toDelete.getId() != null) {
+			Session session = sessionFactory.getCurrentSession();
+			Transaction tx = session.beginTransaction();
+			try{
+				session.delete(toDelete);
+				tx.commit();
+			}catch(HibernateException e){
+				e.printStackTrace();
+				toDelete = null;
+				tx.rollback();
+			}
+		}else{
 			toDelete = null;
-			tx.rollback();
 		}
 		return toDelete;
 	}
